Rename file scss
