+++ src/components/Book/index.tsx
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/router'
-import styles from './styles.module.scss'
+import styles from './index.module.scss'
 
 export type BookProps = {
   imageURL: string
