+++ src/components/Layout/Footer/index.module.scss
